Had to delete the old repository and remake it
Because I made a grave mistake and didn't know, that I cannot delete pull requests once they are created, I used my personal account to test, that I can create pull requests on this repository which made my name visibly and connectable to the writings. But as I want to stay anonymous, I had to take drastic measures in deleting and remaking the repo.